+++ NEWS.md
@@ -1,3 +1,9 @@
+# MazamaSpatialUtils 0.3.0
+
+Version 0.3.x is an upgrade to accommodate **tmap 4.0**. Some arguments to 
+`stateMap()` and `countyMap()` have changed to reflect changes in the **tmap**
+package.
+
 # MazamaSpatialUtils 0.2.0
 
 Version 0.2.x is a refactoring to accommodate **MazamaSpatialUtils** 0.8 which